+++ src/check_dependencies/app_config.py
@@ -201,12 +201,6 @@ class AppConfig:
             Toml Key: [tool.check-dependencies] includes=[]
             """),
         )
-        parser.add_argument(
-            "--all",
-            action="store_true",
-            help="(Deprecated) Show all imports (including correct ones). "
-            "Use --output-format full.",
-        )
         full, concise, github = (
             OutputFormat.FULL.value,
             OutputFormat.CONCISE.value,
@@ -227,9 +221,6 @@ class AppConfig:
             default=OutputFormat.CONCISE,
         )
         args = parser.parse_args(sysv)
-        if args.all and args.output_format == OutputFormat.CONCISE:
-            logger.warning("--all is deprecated, use --output-format full instead.")
-            args.output_format = OutputFormat.FULL
 
         return AppConfig.from_cli_args(
             file_names=args.file_name,